테이블구조에 맞춰 수정함.
