--- src/main/java/com/SYVegas/common/SearchCriteria.java
@@ -1,39 +1,0 @@
-package com.SYVegas.common;
-
-public class SearchCriteria {
-
-    private String condition;       // 검색조건
-    private String value;           // 검색어
-
-    public SearchCriteria() {
-    }
-
-    public SearchCriteria(String condition, String value) {
-        this.condition = condition;
-        this.value = value;
-    }
-
-    public String getCondition() {
-        return condition;
-    }
-
-    public void setCondition(String condition) {
-        this.condition = condition;
-    }
-
-    public String getValue() {
-        return value;
-    }
-
-    public void setValue(String value) {
-        this.value = value;
-    }
-
-    @Override
-    public String toString() {
-        return "SearchCriteria{" +
-                "condition='" + condition + '\'' +
-                ", value='" + value + '\'' +
-                '}';
-    }
-}
