--- src/main/java/com/SYVegas/customerMyInfo/CustomerMyInfo.java
@@ -1,12 +1,0 @@
-package com.SYVegas.customerMyInfo;
-
-import com.SYVegas.common.CustomerDTO;
-import com.SYVegas.common.SearchCriteria;
-
-import java.util.List;
-
-public interface CustomerMyInfo {
-
-
-    List<CustomerDTO> searchCustomer(SearchCriteria searchCriteria);
-}
