+++ src/main/java/com/SYVegas/customerMyInfo/CustomerMyInfoMapper.java
@@ -1,0 +1,5 @@
+package com.SYVegas.customerMyInfo;
+
+public interface CustomerMyInfoMapper {
+    CustomerDTO selectCustomerById(String id);
+}
