+++ src/main/java/com/SYVegas/customerMyInfo/testapplication.java
@@ -1,34 +1,32 @@
 package com.SYVegas.customerMyInfo;
 
-import com.SYVegas.common.SearchCriteria;
+import com.SYVegas.common.Template;
+import org.apache.ibatis.session.SqlSession;
 
 import java.util.Scanner;
 
 public class testapplication {
 
     public static void main(String[] args) {
-        Scanner sc = new Scanner(System.in);
-        do{
-            System.out.println("1. 검색 ");
-            System.out.println("9. 프로그램 종료");
-            System.out.println("검색조건을 입력해주세요 :");
-            int no = sc.nextInt();
-            switch (no){
-                case 1: CustomerInfoRead.searchCustomer(inputSearchCriteria());break;
-                case 9:
-                    System.out.println("프로그램을 종료합니다. "); return;
-            }
-        }while(true);
+        Scanner scanner = new Scanner(System.in);
+        System.out.print("검색할 고객의 ID를 입력하세요: ");
+        String id = scanner.nextLine();
+
+        CustomerDTO customer = searchCustomerById(id);
+        if (customer != null) {
+            System.out.println(customer);
+        } else {
+            System.out.println("일치하는 고객 정보가 없습니다.");
+        }
     }
-
-    private static SearchCriteria inputSearchCriteria() {
-        Scanner sc = new Scanner(System.in);
-        System.out.println("검색 기준을 입력하세요 (id 또는 name) : ");
-        String condition = sc.nextLine();
-        System.out.println("검색어를 입력하세요 : ");
-        String value = sc.nextLine();
-
-        return new SearchCriteria(condition, value);
+    private static CustomerDTO searchCustomerById(String id) {
+        SqlSession sqlSession = Template.getSqlSession();
+        try {
+            CustomerMyInfoMapper mapper = sqlSession.getMapper(CustomerMyInfoMapper.class);
+            return mapper.selectCustomerById(id);
+        } finally {
+            sqlSession.close();
+        }
     }
-
 }
+
