--- src/main/resources/com/SYVegas/customerMyInfo/CustomerMyInfo.xml
@@ -1,32 +1,0 @@
-<?xml version="1.0" encoding="UTF-8" ?>
-<!DOCTYPE mapper
-        PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN"
-        "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
-<mapper namespace="com.SYVegas.customerMyInfo.CustomerMyInfo">
-    <resultMap id="customerResultMap" type="com.SYVegas.common.CustomerDTO">
-        <id property="id" column="customer_id"/>
-        <result property="name" column="customer_name"/>
-        <result property="password" column="customer_password"/>
-        <result property="phone" column="customer_phone"/>
-        <result property="birth" column="customer_birth"/>
-        <result property="address" column="customer_address"/>
-        <result property="balance" column="customer_balance"/>
-        <result property="credit" column="customer_credit"/>
-        <result property="rank" column="customer_rank"/>
-    </resultMap>
-
-    <select id="searchCustomer" parameterType="com.SYVegas.common.SearchCriteria" resultMap="customerResultMap">
-        SELECT customer_id, customer_name, customer_password, customer_phone,
-        customer_birth, customer_address, customer_balance,
-        customer_credit, customer_rank
-        FROM Customer
-        WHERE 1=1
-        <if test="condition == 'id'">
-            AND customer_id LIKE #{value}
-        </if>
-        <if test="condition == 'name'">
-            AND customer_name LIKE CONCAT('%', #{value}, '%')
-        </if>
-        ORDER BY customer_id
-    </select>
-</mapper>
