+++ src/main/resources/com/SYVegas/customerMyInfo/CustomerMyInfoMapper.xml
@@ -1,0 +1,11 @@
+<?xml version="1.0" encoding="UTF-8" ?>
+<!DOCTYPE mapper PUBLIC "-//mybatis.org//DTD Mapper 3.0//EN" "https://mybatis.org/dtd/mybatis-3-mapper.dtd">
+<mapper namespace="com.SYVegas.customerMyInfo.CustomerMyInfoMapper">
+    <select id="selectCustomerById" parameterType="String" resultType="com.SYVegas.common.CustomerDTO">
+        SELECT customer_id, customer_name, customer_password, customer_phone,
+        customer_birth, customer_address, customer_balance,
+        customer_credit, customer_rank
+        FROM Customer
+        WHERE customer_id = #{id}
+    </select>
+</mapper>
