+++ src/main/resources/config/mybatis-config.xml
@@ -4,7 +4,9 @@
         "https://mybatis.org/dtd/mybatis-3-config.dtd">
 <configuration>
     <properties resource="config/connection-info.properties"/>
-
+    <typeAliases>
+        <typeAlias alias="CustomerDTO" type="com.SYVegas.customerMyInfo.CustomerDTO"/>
+    </typeAliases>
     <environments default="dev">
         <environment id="dev">
             <transactionManager type="JDBC"/>
@@ -17,8 +19,7 @@
         </environment>
     </environments>
     <mappers>
-        <mapper resource="com/SYVegas/customerMyInfo"/>
+        <mapper resource="com.SYVegas.customerMyInfoMapper"/>
     </mappers>
-
 
 </configuration>
